pyproject.toml
[tool.poetry]
name = "Trip Q"
version = "0.1.0"
description = "A Django REST API for a local bus route planning system"
authors = ["Sasindu Sulochana <sasindusulochana99@gmail.com>"]
license = "MIT"
readme = "README.rst"
packages = [
    { include = "vvecon" },
]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
daphne = "^4.1.2"
pytest = "^8.2.1"
pytest-xdist = "^3.6.1"
pytest-django = "^4.8.0"
model-bakery = "^1.18.0"
python-dotenv = "^1.0.1"
setuptools = "^70.1.1"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.1"
colorlog = "^6.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
